+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Docs
 
 ### Under the hood
+* Removed `npm logout` from `publish-nodejsdriver` gulp task as it was destroying npmjs access tokens https://stackoverflow.com/questions/68929900/npm-read-only-tokens-keep-getting-deleted